refactor: Header 코드 분리 및 menu 추가
